let's split out the data prep stuff
